Create GraphQL configuration file
The command
`php artisan vendor:publish --provider="Rebing\GraphQL\GraphQLServiceProvider"`
extracts the graphql.php configuration file from the vendor
folder to the project's config folder. This is a common
approach that allows us to get config files from a third
party package so we can easily change it for the need of
our application.