Create API middleware
Middleware to check if the HTTP requests sent to the API
are authorized. Created middleware using php artisan
make:middleware command.

The handle method checks the access token only if there is
a configured Auth0 audience and domain. If that's the case,
this method creates an instance of the JWTVerifier class
based on Auth0 configuration data sen in the .env file. Then
this method retrieves the current bearer token from the
HTTP request and verifies the JWTVerifier instance.

If the token is valid, the request is forwarded to GraphQL.
Access denied otherwise.